+++ CHANGELOG.rst
@@ -30,6 +30,8 @@ New Features
 Breaking changes
 ^^^^^^^^^^^^^^^^
 
+- Localization radii that lead to singular matrices are now skipped (`#187 <https://github.com/MESMER-group/mesmer/issues/187>`__).
+  By `Mathias Hauser <https://github.com/mathause>`_.
 - Refactor and split :py:func:`train_l_prepare_X_y_wgteq` into two functions:
   :py:func:`get_scenario_weights` and :py:func:`stack_predictors_and_targets`
   (`#143 <https://github.com/MESMER-group/mesmer/pull/143>`_).
@@ -67,9 +69,13 @@ Internal Changes
   By `Mathias Hauser <https://github.com/mathause>`_.
 - Move contents of setup.py to setup.cfg (`#169 <https://github.com/MESMER-group/mesmer/pull/169>`_).
   By `Mathias Hauser <https://github.com/mathause>`_.
+- Refactored functions dealing with the spatial covariance and its localization (`#167 <https://github.com/MESMER-group/mesmer/pull/167>`__
+  and `#184 <https://github.com/MESMER-group/mesmer/pull/184>`__).
+  By `Mathias Hauser <https://github.com/mathause>`_.
 - Use pyproject.toml for the build-system and setuptools_scm for the `__version__`
   (`#188 <https://github.com/MESMER-group/mesmer/pull/188>`_).
   By `Mathias Hauser <https://github.com/mathause>`_.
+
 
 v0.8.3 - 2021-12-23
 -------------------